ADDING MODEL GETFAVORITEBYID, AND UPDATING CREATE FAVORITE CONTROLLER WITH THE NEW MODEL

+++ src/controllers/favoriteController.js
@@ -2,6 +2,7 @@ const {
   getFavorites,
   newFavorite,
   deleteFavorite,
+  getFavoriteById,
 } = require("../models/favorites");
 
 const all = async (req, res) => {
@@ -15,6 +16,15 @@ const all = async (req, res) => {
 };
 const create = async (req, res) => {
   try {
+    const existingFavorite = await getFavoriteById(
+      req.params.user_id,
+      req.body.product_id
+    );
+    if (existingFavorite) {
+      return res
+        .status(409)
+        .json({ error: "Ya existe un favorito con este ID" });
+    }
     const favorite = await newFavorite(req.body.product_id, req.params.id);
     res.status(201).json(favorite);
   } catch (error) {
@@ -24,8 +34,14 @@ const create = async (req, res) => {
 };
 const destroy = async (req, res) => {
   try {
-    const favorite = await deleteFavorite(req.body.product_id, req.params.id);
-    res.status(200).json(favorite);
+    const existingFavorite = await getFavoriteById(
+      req.params.user_id,
+      req.body.product_id
+    );
+    if (existingFavorite) {
+      const favorite = await deleteFavorite(req.body.product_id, req.params.id);
+      res.status(200).json(favorite);
+    }
   } catch (error) {
     console.error("Error al remover la obra de favoritos:", error);
     res.status(500).json({ error: error.message });
